Comment updates. Moved Request to its own folder
